Create Task Category Button

+++ employeemanagementsystem/src/main/java/com/employeemanagementsystem/employeemanagementsystem/TaskData.java
@@ -5,14 +5,14 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class TaskData {
-    public static Map<String, ArrayList<Long>> taskDataLists = new HashMap<>();
+    public static Map<String, ArrayList<Double>> taskDataLists = new HashMap<>();
 
     public static void createTaskDataArray(String taskName) {
-        ArrayList<Long> taskDataList = new ArrayList<>();
+        ArrayList<Double> taskDataList = new ArrayList<>();
         taskDataLists.put(taskName, taskDataList);
     }
 
-    public static ArrayList<Long> gettaskDataList(String taskName) {
+    public static ArrayList<Double> gettaskDataList(String taskName) {
         return taskDataLists.get(taskName);
     }
 }
